requirements.txt
ansi2html==1.9.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
nest-asyncio==1.5.9
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
psycopg2-binary==2.9.9
statsmodels==0.14.1